Jetpack Settings: Temporarily move utils to jetpack-onboarding dir
